stock_dashboard_lite/requirements.txt
streamlit
yfinance
pandas
matplotlib
python-pptx
textblob